requirements.txt
django>=5.0,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
stripe>=7.0.0
python-decouple>=3.8
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
whitenoise>=6.6.0

